Header for input file
Optionally, the input file may have a header, and in that case, the
format parameters will have to be a reference to the name of the
specific fields of the header. The header support is available for
bifixer as well as monofixer.

Moved mandatory parameters to the predefined group of parameters.